Add cr_correction_factor = 0.99 to avoid dispersion estimates of +Inf
The problem is that for some combinations of y, mu, and X the term
lgamma(1/theta) and the log(det(t(X) %*% W %*% X)) with
W = diag(1/(1/mu + theta)) canceled each other exactly out for large
theta.

To avoid is I break that symmetry and always only add cr_term * 0.99
For most cases this does not affect the result a lot, so it seems to be
a fair trade-off between handling the edge case better and not affecting
the standard performance too badly.