Add a diagonal fisher scoring approximation
It is faster for design matrices with many columns, because it can
be slow to calculate the full info matrix = Xt W X. Instead one
can calculate the diagonal elements of Xt W X linear with the number
of columns. The exact cost of the approximation depend on the
correlation structure of the columns. If the columns are all independent
then fitBeta_diagonal_fisher_scoring would always be better, if they
are not than it is a balance between additional precision from the
off-diagonal elements and the runtime of O(p) vs O(p^2)