Introduce a mu_rel_change parameter to Fisher scoring
This new parameter stops the algorithm from occasionally converging to
a bad result, where the Fisher scoring step jumps off to an extreme
value (e.g. mu = 1e50), but still manages to decrease the deviance.

This is particularly bad for testing: if the algorithm is not converging
for H0, the likelihood ratio test will indicate a large change
of deviance and thus a very significant result, regardless of the true
significance.