Adapt test to new SummarizedExperiment() requirement
Starting with SummarizedExperiment 1.23.2, the SummarizedExperiment()
constructor raises an error if one of the supplied assays has rownames
and/or colnames that don't match those inferred from the supplied 'rowData'
and 'colData' arguments. This change breaks a test in glmGamPoi ("glm_gp
warns about mismatching col_data rownames") and this commit adapt the test
to make it work with the new SummarizedExperiment() behavior.